# docs/swagger.yaml
consumes:
- application/json
definitions:
...
  http.StepResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      orderNo:
        type: integer
      planId:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  http.StepUpsertRequest:
    properties:
      description:
        example: Pause subscriptions and renegotiate vendor costs
        type: string
      orderNo:
        example: 1
        type: integer
      title:
        example: Cut non-essential expenses
        type: string
    required:
    - orderNo
    - title
    type: object
...
info:
  contact: {}
  description: API for Business Surgery service.
  title: Business Surgery Backend API
  version: "1.0"
paths:
...
  /api/v1/plans/{planId}/steps:
    get:
      parameters:
      - description: Plan ID (uuid)
        in: path
        name: planId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.StepResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List steps by plan id
      tags:
      - steps
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer {accessToken}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Plan ID (uuid)
        in: path
        name: planId
        required: true
        type: string
      - description: Step payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.StepUpsertRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.StepResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add step to plan
      tags:
      - plans
  /api/v1/plans/{planId}/steps/{stepId}:
    delete:
      parameters:
      - description: Bearer {accessToken}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Plan ID (uuid)
        in: path
        name: planId
        required: true
        type: string
      - description: Step ID (uuid)
        in: path
        name: stepId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete step
      tags:
      - plans
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer {accessToken}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Plan ID (uuid)
        in: path
        name: planId
        required: true
        type: string
      - description: Step ID (uuid)
        in: path
        name: stepId
        required: true
        type: string
      - description: Step payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.StepUpsertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.StepResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update step
      tags:
      - plans
produces:
- application/json
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"